refactor: simplify codebase by extracting common patterns
Reduce duplication and improve maintainability across TypeScript plugins,
emitters, and test files without changing functionality:

**Test utilities** (~120 lines saved):
- Extract `removePositions()` and `createRehypeProcessor()` to shared test-utils.ts
- Used by spoiler, subtitles, and tablecaption tests

**Transformer plugin factory** (~50 lines saved):
- Create `createElementVisitorPlugin()` factory for common visitor pattern
- Refactor spoiler, subtitles, and tablecaption plugins to use factory
- Eliminates boilerplate plugin wrapper code

**Emitter page factory** (~140 lines saved):
- Create `createListPageEmitter()` factory for list page generation
- Consolidate allPostsPage and allTagsPage (95% identical) into single pattern
- Both files now <15 lines vs original ~80 lines each

**Type guards consolidation**:
- Move `isTextNode()`, `isElementNode()`, `isCode()` to utils.ts
- Re-export from tablecaption.ts for backward compatibility
- Update formatting_improvement_html.ts and tagSmallcaps.ts imports

**Total impact**: ~310 lines removed, improved code organization and reusability
**Test coverage**: Maintained 100% branch/statement/function/line coverage
**All tests pass**: 3223 tests, 59 suites
